完善 todo 部分的开发，优化前端部分的表现效果，整体开发完成，添加初试 sql 文件

+++ model/dao/todo_dao.go
@@ -15,7 +15,7 @@ func GetTodoById(id uint64, userID uint64) (*model.Todo, error) {
 	return &todo, err
 }
 
-func GetTodoList(userID uint64, status *int, offset, limit int) ([]model.Todo, int64, error) {
+func GetTodoList(userID uint64, status *int8, offset, limit int) ([]model.Todo, int64, error) {
 	var todos []model.Todo
 	var total int64
